+++ src/main/java/com/daitda/hubservice/domain/hubinventory/application/dto/DecreaseHubInventoryCommand.java
@@ -1,4 +1,4 @@
-package com.daitda.hubservice.hubinventory.dto.command;
+package com.daitda.hubservice.domain.hubinventory.application.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 @Getter
 @Builder
-public class RestoreHubInventoryCommand {
+public class DecreaseHubInventoryCommand {
 
     private UUID hubInventoryId;
     private Integer quantity;
